Adding an script that write particle information parallelized over grids
